requirements.txt
streamlit==1.31.0
langgraph==0.0.19
langchain==0.0.335
langchain-community==0.0.20
langchain-openai==0.0.5
openai==1.6.1
chromadb==0.4.22
duckduckgo-search==3.9.9
pydantic==2.5.2
pyyaml==6.0.1
python-dotenv==1.0.0